requirements.txt
Pillow>=9.0.0
requests>=23.2.1